incremental-deploy: deploy aggregate POMs more intelligently
For aggregate POMs, there is no way to determine the commit from which
they were deployed. However, we can be clever about it nevertheless:
they only need to be deployed when at least one of their children
was deployed.

While at it, we explicitly only handle 'jar' and 'pom' packaging
(and not, say, 'nar').

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>