bump-pom-scijava: add special exit status for Jenkins
In the Jenkins job, we will want to update other projects *iff* the
pom-scijava was updated. So we need to discern between 'there was an
error' and 'no commit needed'.

GNU suggests using exit status 128 when a subprocess failed to execute,
and in a certain sense, this is kind of true: if no new version was
committed, the commit failed to execute...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>