release-version.sh: support extra arguments
The maven-release-plugin lets you pass extra arguments to child
executions using the '-Darguments=' flag.

In my case, I need it to pass '-Dgpg.skip=true' so that we can cut
custom releases of third-party software (specifically nar-maven-plugin)
without signing the release.

To support this, add the --extra-args and --skip-gpg options.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>