release-version.sh: add --skip-deploy flag
We need this to release the ImageJ launcher, because its deployment is
more complicated. It needs to build on several architectures and then
deploy them all simultaneously. We already have our trusty Jenkins for
this, so let's skip the deploy when running the script, and instead tell
Jenkins to do the deploy from the tag after it is pushed.