release-version.sh: verify that maven-release-plugin did its job
Under circumstances yet to be determined, maven-release-plugin fails to
adjust the <version> tag to the correct release version, then fails to
notice that and make an incorrect tag (still referencing the SNAPSHOT
version), and finally fails to commit and push changes bumping to the next
release cycle.

Detect that situation and fail.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>