bump-pom-scijava: after deploying to OSS Sonatype, invalidate cache
Our primary mirror is the ImageJ Maven repository. This also happens to be
the go-to place for the maven-helper when looking for the latest version.

Therefore, let's just use the RESTful API of the Nexus instance that is
the ImageJ Maven repository to invalidate the cache of maven-metadata.xml
just after releasing to OSS Sonatype.

That way, the maven-helper will trigger that file to be fetched from OSS
Sonatype, merge it with what we have in the local Releases repository
(that was formerly the official place where we deployed pom-scijava), and
everything is groovy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>