Move the code to invalidate the cached pom-scijava to maven-helper
We will need this functionality a lot. A *lot*. Because we will release
more and more stuff to OSS Sonatype and will require maven-helper to
help us with adjusting things afterwards, for which maven-helper will
need up-to-date cached versions, of course.

Since this code was adjusted from bump-pom-scijava, just use the new
'invalidate-cache' sub-command there, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>