sj-hierarchy.pl: process snapshot artifacts, too
If an artifact has never had a release, maven-helper.sh reports its
latest version as a SNAPSHOT. We do not want to exclude this situation
from our analysis; on the contrary, we want to test compilation of these
artifacts against everything else, too.

This change also adds obsolete SNAPSHOT-only artifacts to the blacklist.