sj-hierarchy.pl: use good version of dep plugin
The maven-dependency-plugin at version 2.8 (latest version as of this
writing) works just fine with the parameters used by this script.
But older versions of Maven core (3.0.4 in my test case) use a much
older version of maven-dependency-plugin, 2.1, which does *not* work.
So, rather than worrying about needing a minimum version of Maven core
(which we don't, really), let's just force maven-dependency-plugin 2.8.