bump-pom-scijava: change property range
As the pom-scijava pom.xml now contains multiple properties ranges, it
is not enough to search and replace within the properties tag.

To catch only the first properties section, we now rely on a properly
indented pom-scijava pom.xml, as we harden the sed range by making the
single tab character part of the pattern to match the appropriate section.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>