maven-helper: Fix latest-version for SNAPSHOT versions on Sonatype
Sonatype does not store any <release> or <latest> tags in its
maven-metadata.xml files.

Our fall-back is to extract the first <version> tag and use its value
instead, but that is incorrect: it needs to be the *last* <version>
tag.

This should fix the bump-pom-scijava functionality that is intended
to update the snapshot versions.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>