sj-hierarchy.pl: ignore artifacts without releases
For now, it is less complicated to stop processing SNAPSHOT versions
altogether. Otherwise, we have to take care to update the SNAPSHOT
metadata in ~/.m2/repository every 24 hours, and include other
snapshot-specific case logic in various other places as well.

In the future we could restore handling of snapshots, but further
thought would be needed to make it work in a sensible way.