sj-hierarchy.pl: do not include ome artifacts
The ome groupId only applies to Bio-Formats, and the Bio-Formats
project does not even use pom-scijava as parent anymore.

Further, the build is quite slow, and not 100% pure Java. So for the
main intended use case of this script -- generating a big melting pot
to build all at once -- it does not make sense to include Bio-Formats.