verify-checksums.sh: support alt. checksum format
Some checksum files do not simply have the checksum, but are also
appended by a path; e.g.:

$ cat ~/.m2/repository/ant/ant/1.6/ant-1.6.pom.sha1
04ebfa03f2baf244c7c7bc2a754cb03221ebfc29  /home/projects/maven/repository-staging/to-ibiblio/maven2/ant/ant/1.6/ant-1.6.pom

As an aside, this is almost certainly an error in the version of the
hashing plugin used to write that file -- the default output of sha1sum
is to append the full path to the file after the checksum, so presumably
that information was not removed as it seems to be with newer artifacts.

In any case, it's easy for this script to support both formats.