release-version.sh: update for the ImgLib2 split
The ImgLib2 project is no longer a multi-module build, but now lives in
separate Git repositories as individual components, analogous to the
SciJava, SCIFIO, ImageJ and Fiji organizations.

The pom-imglib2 parent POM will be released to OSS Sonatype. The
net.imglib2:imglib2 artifact will be released to Sonatype as well.
The other imglib2-* artifacts will continue to be released to the
ImageJ Maven repository for the time being, until they come out of beta.