check-release.pl: add hack for underscore suffixes
In the Fiji project are many ImageJ 1.x plugins. Each such plugin must
have an underscore somewhere in the name; and when there are not
multiple words in the artifactId, the underscore is typically appended
to the end. But for aesthetic reasons, the repository name and tags have
that underscore stripped. So let's strip it here too, since we have no
artifacts that "legitimately" end with an underscore.