check-release.pl: fix release version detection
When gleaning the latest release from the Maven metadata file, the grep
system call (like all system calls from Perl) has a trailing newline.
But we do not want to return the version string to include that newline.

This bug was likely the cause of some weird behavior in
status.imagej.net at various points, including misreporting a component
as being dozens or hundreds of commits ahead of the last release.