release-version.sh: do not strip pom- prefixes
We used to have lots of multimodule projects with the root aggregator
called "pom-foo", and we wanted it tagged as "foo-x.y.z". But now, all
that has changed: the only project still like that is TrakEM2. The other
projects are all single modules, released independently, including the
pom-* parents like pom-scijava, pom-imagej, pom-fiji, etc. So in
general, let's not strip the pom- prefix, but instead only do it for
components we know encapsulate an entire multi-component project.