melting-pot: add flag to prune unaffected deps
The default behavior of the melting pot is to create a multi-module
build of all project components which match the inclusion criteria
(i.e., the includes and excludes), minus any changed components
specified by the '--changes' flag.

However, when the set of changes is small, and the main interest is in
testing breakages relating to those changes, then it is nice to limit
the build to only those components which depend on the changed
components (either directly or transitively).

Hence, the new '--prune' flag does exactly that!

Be warned that even if a component does not explicitly depend on a
changed component, it might still be affected by changes in that
component. One example is components which use dependency injection at
runtime, using resources discovered on the classpath (e.g., plugins).