melting-pot: add option to override the branch
For projects like sc.fiji:fiji:2.0.0-SNAPSHOT, there is no tag
or branch called 'fiji-2.0.0-SNAPSHOT' because it is a snapshot.

The -b option allows to specify the branch to use explicitly;
e.g., for the fiji GAV above, it should be master.

So now the following invocation is much closer to working:

  melting-pot.sh sc.fiji:fiji:2.0.0-SNAPSHOT -b master \
    -i 'org.scijava:*,net.imagej:*,net.imglib2:*,io.scif:*,sc.fiji:*' \
    -r http://maven.imagej.net/content/groups/public -v -f