check-release.pl: also list the commit total
When filtering the commits to only those relevant, it is still nice to
know the total number of commits since the last release, even if some of
them got filtered out. So let's list the commits as "x/y" where x is the
number of relevant commits, and y is the total since last release.

Requested by Mark Hiner.