Remove incremental deploy script
Nothing uses it anymore.

It remains a potential issue that deploying an unchanged SNAPSHOT is
potentially wasteful in both time and space.

But actually, we no longer do daily builds of anything -- we only build
components in response to triggers, particularly pushes to master.
So the assumption is that the new artifact will typically differ.