Purge no-longer-needed verify-code-snippets script
This script was used to synchronize code between GitHub and web pages,
including MediaWiki and Drupal sites. But we no longer need it.

Instead, we now have this MediaWiki extension:

  https://github.com/imagej/mediawiki-GitHub

Unfortunately, it is a fork of the original project,
but the changes are quite small:

  https://github.com/JeroenDeDauw/GitHub/compare/master...imagej:master

And it works great for what the ImageJ wiki needs.

I know of no other uses of this script in the wilds of the SciJava
ecosystem, so I am boldly retiring it. Complain loudly if you need it.