release-version: check for .netrc credentials
If they are not present, then the script will not be able to
invalidate the remote cache after deploying the release.

Currently, we use an account called "jenkins-expire-cache" to do this,
which is now a misnomer, and also pretty hacky. It would be better to
simply use the current user's credentials for doing it, and ensure
that that user has the relevant permission on the server side.
But I don't have time to research how to do that at the moment.