release-version.sh: remove hardcoded GAVs
Instead, we check the value of the releaseProfiles property in the POM;
if it equals "sonatype-oss-release" then we enable OSS Sonatype stuff.

This way, each project can decide for itself whether to be released
to the ImageJ Maven repository, OSS Sonatype, or somewhere else.

It also removes case logic for the ImageJ Launcher (which needs to
have its release process tweaked regardless), TrakEM2 (which really
cannot be properly released with this script anyway), and
jython-shaded (which now overrides the tagNameFormat property to
hopefully achieve the same result of stripping the pom- prefix).