release-version.sh: add release.properties to tag
The release tag will be used by Travis CI remotely to complete the
release via the release:perform goal. That goal is helped along very
much by a release.properties file. This file contains no secrets, and is
handy for performing releases, so we add it to the tag for convenience.