release-version.sh: add hack to fix remote release
The idea of committing the release.properties was to allow a different
machine to perform release:perform than performed release:prepare.

However, release:perform tries to check out the release tag using the
developerConnection, which we have set to git@github.com:... since
that is indeed an appropriate read/write SCM connection URL.

If the machine performing release:perform (e.g., a Travis CI instance)
does not have SSH credentials to the repository, the script will fail.
So as a workaround, we replace the scm.url with the read-friendly
regular connection URL (git://github.com/...), since release:perform
does not need push access to the SCM in order to complete the release.