release-version: remove unneeded sed command
GNU and BSD sed commands are slightly different. In particular, the GNU
sed does not like a space after the -i flag. I believe it used to be the
case that BSD sed refused to work _without_ that space, but in my tests
today, it worked with -i'' the same as the GNU version.

But all of that is moot because it seems that this sed command was never
actually doing anything anyway. At least, examining the
imagej-common-0.24.8 tag (from imagej/imagej-common) as well as
Correct_3D_Drift-1.0.2 (from fiji/Correct_3D_Drift), it seems that both
release.properties files retained git@github.com developerConnection,
without ill effects during release, so... never mind.

Thanks to Hadrien Mary for pointing out this issue.

For related discussion, see:
https://gitter.im/scijava/scijava-common?at=59e14ea65c40c1ba799a9641