travis-build.sh: bootstrap settings.xml internally
Rather than relying on every repository always having its
own (in all cases so far identical) .travis/settings.xml,
the script can generate one itself for its own use.

If a .travis/settings.xml does exist, however, we use that.
This way, individual repositories can still customize the
settings if they need to do so for some reason -- although
doing so runs the risk of getting "out of sync" with this
script's requirements, if they change in the future.