travis-build.sh: enable GPG signing if feasible
If signingkey.asc.enc is present, decrypt it and import it into the
GPG keyring. On the Maven side, if ~/.gnupg exists, activate a profile
setting the gpg.keyname & gpg.passphrase Maven properties to match the
GPG_KEY_NAME and GPG_PASSPHRASE environment variables, respectively.

These changes pave the way for signing build artifacts as needed;
see also the forthcoming add-travis.sh script in this repository.