travis-build.sh: import GPG key, but don't decrypt
The decryption will be done in the before_install phase, as
specified in the .travis.yml. This is what "travis encrypt-file"
sets up when you pass "--add". It is a nice way to do it because
the environment-specific stuff all stays in .travis.yml at the end,
and we do not need to pass key and iv values into travis-build.sh.