travis-build.sh: trigger Maven build by pom.xml
If a pom.xml is not present, there is no need to do all the Maven stuff.

This change will pave the way for this script to become more general,
with additional and/or alternative steps besides only Maven builds.

You may want to view this diff with the "-w" flag.