travis-build.sh: add conda installation + config
That way, everyone does not need their own complex installation step
for conda. Instead, the build script takes care of it, in the same way
it takes care of configuring things like GPG for Maven.

We do it in such a way that if the ~/miniconda installation folder is
cached from previous builds, we can hopefully gracefully update it,
rather than needing to download and install a new miniconda every time.