travis-javadoc: make output more verbose
We now prefix the various steps explaining what is going on, similar
to the travis-build.sh script. And most importantly, we launch the
quiet javadoc command using travis_wait, so that it emits a small
amount of output each minute, to avoid the "No output has been
received in the last 10m0s" timeouts. We do not simply remove the
-q flag, because it would make the javadoc build log enormous.