release-version.sh: remove obsolete GPG check
It looked for <releaseProfiles>sonatype-oss-release</releaseProfiles>
in the <properties>. But the default for SciJava projects is now
sonatype-oss-release, meaning if the releaseProfiles property is not
explicitly set, it will try to release to Sonatype. So the test was
not flagging projects configured to release to OSS Sonatype anyway.

It is true that .travis/signingkey.asc.enc should be present for
projects intended for release to OSS Sonatype, and if it is not, then
the project will almost certainly fail to be released as intended.
The release_version.sh script could check for this, but it does not
check for similar prerequisites such as encrypted environment variable
entries in the .travis.yml. So for now, let's not bother checking it.