github-actionify.sh: fix sed expressions
The repoSlug contains a slash, which will affect the expression
if slash is used as the delimiter. So we change to semicolon.

The command "sed -i" with no extension does not work with BSD sed
(e.g. macOS); you need to write "sed -i ''" to avoid saving the
previous file to a backup. But this doesn't work with GNU sed,
which requires "sed -i''" or simply "sed -i" instead. The only
way I know that works with both sed flavors is to give the flag
as "-i.bak" and then delete the .bak file afterward.