github-actionify: improve var names and filenames
There is no reason for the YAML files to begin with a dot character.
These are not hidden files; they are just build configuration files.

The workflows in question are the main build, and the PR build, so we
now name them as such. While a project's CI may have other workflows
besides these, in the standard case, these two workflows are the two
used by the vast majority of SciJava projects. Projects needing
customization beyond this are of course free to rename everything
however they see fit, hack the build scripts, or do whatever.