ci-build.sh: change how settings.xml is generated
It seems that settings.xml sometimes(?) already exists on GitHub Actions
infrastructure. Either that, or it is/was getting cached between runs
(although that is not intended). Regardless: builds are now failing to
deploy, because the already-existing settings.xml does not contain the
proper credentials, and a correct one is no longer being written out.

This commit changes the script to overwrite settings.xml when either
MAVEN_PASS or OSSRH_PASS is set, even if ~/.m2/settings.xml already
exists. On a typical developer workstation, these variables won't be
set, so running ci-build.sh locally will not stomp ~/.m2/settings.xml.