release-version.sh: try again with -U as needed
It's nice if, when failing to extract the version due to dependencies
not being present in the remote repository, we first try again using -U
before giving up. Otherwise, we may end up with the unfortunate error:

  Failure to find <artifact> in
  https://maven.scijava.org/content/groups/public was cached in the
  local repository, resolution will not be reattempted until the update
  interval of scijava.public has elapsed or updates are forced

If the artifact has been successfully deployed after the local system
last checked for it, but 24 hours has not yet elapsed, the -U flag will
force a new check to find the artifact, and the release can continue.