melting-pot.sh: set long version properties, too
There are sometimes components whose artifactIds are the same, but have
different groupIds. The convention for such components is to set a fully
qualified version property formatted "groupId.artifactId" to avoid the
ambiguity. This change respects that convention by setting both the long
and short version properties accordingly. As long as no two components
in the melting pot use the same short version string for different GAs,
there will be no problem with pinning to the wrong versions.

Closes #34.