ci-build.sh: fail fast on bad version strings
It is incorrect to push a release version string to the main branch
directly, such that the CI deploys it directly. Rather, the
release-version.sh script should be used to cut releases.
The build script now checks for this situation where the version
string is a release version but release.properties is missing,
and fails the build immediately.