melting-pot: improve local git repository caching
This cuts down on excessive network use. The first time a particular G:A
is requested at any version, it gets fully (not shallowly) cloned into
the user's melting-pot cache folder. From there, the tag for the given
version is cloned shallowly from the locally cached repository. In this
way, complex melting pots can be repeatedly generated without hitting
the network in most cases.

This commit also improves debug and info messages in a few places.