ci-build: do not deploy when building PRs
If my reading of the GitHub Actions documentation is correct, then
organization secrets will be available to pull request builds from
a branch of the same repository (as opposed to from a project fork).

In that case, we still want the build script to skip deployment,
because we don't want to be deploying snapshots from from PRs,
only those from the mainline integration branch.