melting-pot: do not generate an aggregator POM
Unfortunately, due to a bug in Maven, this aggregator POM does not work
in an intuitive way. In particular, executing build.sh in the top-level
melting-pot directory does not pin the versions as desired within each
submodule. Shockingly, they often are different versions than when you
cd into a particular groupId/artifactId directory and run ../../build.sh
from there.

The problem is probably related to:
* https://stackoverflow.com/q/45041888/1207769
* https://issues.apache.org/jira/browse/MNG-6141
* https://issues.apache.org/jira/browse/MNG-5761

Anyway, the melting pot does not use this aggregator POM anymore anyway,
so let's refrain from generating it, to avoid any potential reliance on
something that's broken.