release-version.sh: avoid mvn ANSI colors
Firstly, the mvn echo logic needs to pass -B.

But even then, it seems that at least some versions of mvn (3.8.7 is an
offender on my system) still prepend an <ESC>[0m even in batch mode.
So we also work around this bug by forcibly stripping ANSI color codes.

Thanks to ChatGPT for the regex -- saved me a few minutes of time!