Update repository ID for OSSRH -> Central Portal
See:
* https://central.sonatype.org/pages/ossrh-eol/
* https://central.sonatype.org/publish/publish-portal-maven/

For the moment, we leave the environment variable names alone, even
though the prefix "OSSRH" is a misnomer now. Reason: it is possible that
the existing OSSRH user tokens were migrated directly. If they don't
work, I'll change this script again to use new environment variables,
and issue a better error if the new variables are missing.