ci-build.sh: write out settings.xml in pieces
If some env vars are missing, we can still partially write it.
This is especially important if CENTRAL_* vars are missing, but
MAVEN_* vars are present, which as of this writing will be the case
for the vast majority of SciJava-based community projects deploying
to maven.scijava.org, since the transition from OSSRH to Central Portal
is still very recent and many orgs have not yet been updated to match.